+++ src/pylmrob/__init__.py
@@ -13,7 +13,7 @@ from pylmrob.anova import anova
 from pylmrob.bootstrap import bootstrap
 from pylmrob.control import Control
 from pylmrob.lmrob import LmRob, lmrob
-from pylmrob.rng import RState, r_set_seed
+from pylmrob.rng import RState, r_sample_noreplace, r_set_seed
 
 __all__ = [
     "Control",
@@ -24,5 +24,6 @@ __all__ = [
     "bootstrap",
     "lmrob",
     "make_generator",
+    "r_sample_noreplace",
     "r_set_seed",
 ]